Refactor test database fixtures for reuse
Move these to their own respective modules so that they can be
imported by all the tests that need them.
